Organize constants and functions
