+++ anomalous-token-transfers-ml-py/requirements.txt
@@ -1,2 +1,4 @@
 forta_agent>=0.1.1
 setuptools>=61.3.1
+requests>=2.26.0
+scikit-learn>=1.1.1
