+++ anomalous-token-transfers-ml-py/src/constants.py
@@ -1,0 +1,54 @@
+ERC20_TRANSFER_EVENT = '{"name":"Transfer","type":"event","anonymous":false,"inputs":[{"indexed":true,"name":"from","type":"address"},{"indexed":true,"name":"to","type":"address"},{"indexed":false,"name":"value","type":"uint256"}]}'
+
+ETHPLORER_KEY = "BACON"
+ETHPLORER_ENDPOINT = "https://api.ethplorer.io"
+
+LUABASE_ENDPOINT = "https://api.luabase.com/run"
+
+MODEL_FEATURES = [
+    'APE_transfers',
+    'APE_value',
+    'CRV_transfers',
+    'CRV_value',
+    'DAI_transfers',
+    'DAI_value',
+    'GALA_transfers',
+    'GALA_value',
+    'HEX_transfers',
+    'HEX_value',
+    'KOK_transfers',
+    'KOK_value',
+    'LINK_transfers',
+    'LINK_value',
+    'LOOKS_transfers',
+    'LOOKS_value',
+    'MANA_transfers',
+    'MANA_value',
+    'MATIC_transfers',
+    'MATIC_value',
+    'SAITAMA_transfers',
+    'SAITAMA_value',
+    'SAND_transfers',
+    'SAND_value',
+    'SHIB_transfers',
+    'SHIB_value',
+    'SOS_transfers',
+    'SOS_value',
+    'STRNGR_transfers',
+    'STRNGR_value',
+    'STRONG_transfers',
+    'STRONG_value',
+    'USDC_transfers',
+    'USDC_value',
+    'USDT_transfers',
+    'USDT_value',
+    'WBTC_transfers',
+    'WBTC_value',
+    'WETH_transfers',
+    'WETH_value',
+    'account_age_in_minutes',
+    'max_single_token_transfers',
+    'max_single_token_transfers_value',
+    'tokens_type_counts',
+    'transfer_counts']
+
